Move custom built-in scalars into new package
