+++ src/GraphQL.BCLScalars/Types/BigIntGraphType.cs
