+++ src/GraphQL.BCLScalars/Types/ByteGraphType.cs
