+++ src/GraphQL.BCLScalars/Types/DateGraphType.cs
