+++ src/GraphQL.BCLScalars/Types/DateOnlyGraphType.cs
