+++ src/GraphQL.BCLScalars/Types/DecimalGraphType.cs
