+++ src/GraphQL.BCLScalars/Types/GuidGraphType.cs
