+++ src/GraphQL.BCLScalars/Types/SByteGraphType.cs
