+++ src/GraphQL.BCLScalars/Types/ShortGraphType.cs
