+++ src/GraphQL.BCLScalars/Types/TimeOnlyGraphType.cs
