+++ src/GraphQL.BCLScalars/Types/TimeSpanMillisecondsGraphType.cs
