+++ src/GraphQL.BCLScalars/Types/TimeSpanSecondsGraphType.cs
