+++ src/GraphQL.BCLScalars/Types/UIntGraphType.cs
