+++ src/GraphQL.BCLScalars/Types/UShortGraphType.cs
