+++ src/GraphQL.BCLScalars/Types/UriGraphType.cs
