+++ src/GraphQL.Tests/GraphQL.Tests.csproj
@@ -20,6 +20,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\GraphQL.BCLScalars\GraphQL.BCLScalars.csproj" />
     <ProjectReference Include="..\GraphQL.Dummy\GraphQL.Dummy.csproj" />
     <ProjectReference Include="..\GraphQL.MicrosoftDI\GraphQL.MicrosoftDI.csproj" />
     <ProjectReference Include="..\GraphQL.StarWars\GraphQL.StarWars.csproj" />
