+++ src/GraphQL.Tests/Subscription/SubscriptionSchema.cs
@@ -10,6 +10,8 @@ public class ChatSchema : Schema
 {
     public ChatSchema(IChat chat)
     {
+        this.RegisterBCLScalars();
+
         Query = new ChatQuery(chat);
         Mutation = new ChatMutation(chat);
         Subscription = new ChatSubscriptions(chat);
