+++ src/GraphQL.sln
@@ -142,6 +142,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.MicrosoftDI.Tests",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.DataLoader", "GraphQL.DataLoader\GraphQL.DataLoader.csproj", "{466AD883-96BD-458E-9872-3C14E49D18F8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.BCLScalars", "GraphQL.BCLScalars\GraphQL.BCLScalars.csproj", "{97B265CB-7567-4593-ACC4-DAAA35F2FABA}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Scalars", "Scalars", "{E483DF89-D133-41CC-B21F-2B1C5BB0E594}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.BCLScalars.Tests", "GraphQL.BCLScalars.Tests\GraphQL.BCLScalars.Tests.csproj", "{E628DBB1-75BD-4F46-A6CA-F327748CA6AD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -212,6 +218,14 @@ Global
 		{466AD883-96BD-458E-9872-3C14E49D18F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{466AD883-96BD-458E-9872-3C14E49D18F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{466AD883-96BD-458E-9872-3C14E49D18F8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{97B265CB-7567-4593-ACC4-DAAA35F2FABA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{97B265CB-7567-4593-ACC4-DAAA35F2FABA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{97B265CB-7567-4593-ACC4-DAAA35F2FABA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{97B265CB-7567-4593-ACC4-DAAA35F2FABA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E628DBB1-75BD-4F46-A6CA-F327748CA6AD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E628DBB1-75BD-4F46-A6CA-F327748CA6AD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E628DBB1-75BD-4F46-A6CA-F327748CA6AD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E628DBB1-75BD-4F46-A6CA-F327748CA6AD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -232,6 +246,8 @@ Global
 		{92C246A6-E740-41AA-B77A-B939C54B1D8E} = {3DB29B96-B428-46A6-A4D3-FA677110B654}
 		{185923DF-89EE-457B-8DA0-3CBBE080EC32} = {2EA5354E-F4EF-4469-86AE-F1D36C7F8977}
 		{82BE75F5-9C90-427D-8268-1C640A90CE98} = {BC80C071-B8B6-42D2-9CF8-6169C6169B74}
+		{97B265CB-7567-4593-ACC4-DAAA35F2FABA} = {E483DF89-D133-41CC-B21F-2B1C5BB0E594}
+		{E628DBB1-75BD-4F46-A6CA-F327748CA6AD} = {BC80C071-B8B6-42D2-9CF8-6169C6169B74}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {61AAA7F9-9D46-4050-A3A5-F4358BBD95B8}
